+++ lib/src/objects/history.dart
@@ -1,0 +1,16 @@
+class History {
+  DateTime? lastSyncWithNetworkFetchTime;
+
+  History({this.lastSyncWithNetworkFetchTime});
+
+  void updateLastSyncWithNetworkFetchTime() {
+    lastSyncWithNetworkFetchTime = DateTime.now();
+  }
+
+  History copyWith({DateTime? lastSyncWithNetworkFetchTime}) {
+    return History(
+      lastSyncWithNetworkFetchTime:
+          lastSyncWithNetworkFetchTime ?? this.lastSyncWithNetworkFetchTime,
+    );
+  }
+}