+++ lib/src/synced.dart
@@ -6,6 +6,7 @@ import 'package:flutter/services.dart';
 import 'package:synckit/src/utils.dart';
 
 import 'objects/batch.dart';
+import 'objects/history.dart';
 import 'objects/sort_config.dart';
 import 'objects/sync_config.dart';
 
@@ -26,8 +27,11 @@ separate object, which stores data in sorted manner, with log n complexities */
 mixin SyncedState<T> {
   late SyncConfig<T> _params;
   final List<StreamSubscription<Dataset<T>>> _subscriptions = [];
+  final _history = History();
 
   Dataset<T> get state;
+
+  History get history => _history.copyWith();
 
   set state(Dataset<T> value);
 
@@ -58,6 +62,7 @@ mixin SyncedState<T> {
     try {
       final data = await _params.manager.fetchAndSyncFromNetwork;
       _setState(data);
+      _history.updateLastSyncWithNetworkFetchTime();
     } catch (err) {
       log('err: $err', name: 'SyncObjNotifier');
     }
@@ -72,6 +77,7 @@ mixin SyncedState<T> {
   StreamSubscription<Dataset<T>> keepAllInSync() {
     return _addSubscription(_params.manager.listenAllFromNetwork((data) {
       _setState(data);
+      _history.updateLastSyncWithNetworkFetchTime();
     }));
   }
 
@@ -84,7 +90,10 @@ mixin SyncedState<T> {
     return _addSubscription(_params.manager.listenQueryFromNetwork(
       queryFn,
       maxGetAllDocs: maxGetAllDocs,
-      onData: (data) => _updateStateWithDataset(data),
+      onData: (data) {
+        _updateStateWithDataset(data);
+        _history.updateLastSyncWithNetworkFetchTime();
+      },
     ));
   }
 
@@ -108,6 +117,7 @@ mixin SyncedState<T> {
     final data =
         await _params.manager.getQueryFromNetwork(queryFn, maxGetAllDocs);
     _updateStateWithDataset(data);
+    _history.updateLastSyncWithNetworkFetchTime();
     return data;
   }