+++ lib/synckit.dart
@@ -4,6 +4,7 @@ export 'src/local_storage.dart';
 export 'src/network_storage.dart';
 export 'src/objects/batch.dart';
 export 'src/objects/sort_config.dart';
+export 'src/objects/history.dart';
 export 'src/objects/std_obj.dart';
 export 'src/objects/sync_config.dart';
 export 'src/sync_manager.dart';